Duplicate only necessary interfaces and const
